Add files via upload
These are the files I am working with to allow an invite to a developer game & forum.  The first file is a form that just gets an email and encrpyts a confirmation link.  The second file, emailConfirmation, sends the encrypted email link.  Then when the user clicks the link it selects the link from the database and checks it with the link they clicked.  If it is the same (which means its the same person), it goes to approval.php and updates their developer status from 0 to 1.
Still toying with it, not fully functional.